pyproject.toml
[project]
name = "crypto-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = ["requests", "pandas", "python-telegram-bot", "python-dotenv", "notebook"]
